change googlw with login functionality

+++ src/pages/login/index.tsx
@@ -5,7 +5,7 @@ import React, { useState } from "react";
 import * as Yup from "yup";
 import Ic_logo from "@/public/images/Ic_logo.svg";
 import Img_login_google from "@/public/images/Img_login_google.svg";
-import { auth } from "@/config/firebaseConfig";
+import { auth, db } from "@/config/firebaseConfig";
 import Image from "next/image";
 import toast from "react-hot-toast";
 import {
@@ -16,10 +16,14 @@ import {
 import Link from "next/link";
 import { useRouter } from "next/router";
 import Loader from "@/components/Loader";
+import { doc, getDoc, setDoc } from "firebase/firestore";
+import NameModal from "./nameModal";
 
 const Login = () => {
   const router = useRouter();
   const [loading, setLoading] = useState(false);
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [newUser, setNewUser] = useState<any>(null);
 
   const validationSchema = Yup.object().shape({
     email: Yup.string().required("Email is required"),
@@ -44,23 +48,62 @@ const Login = () => {
       router.push("/");
     } catch (error) {
       console.error("Error during sign-in", error);
-      toast.error("Login failed. Please check your credentials.", {
-        position: "top-right",
-      });
+      toast.error(
+        "Login failed. Please check your credentials or register if you don't have an account.",
+        {
+          position: "top-right",
+        }
+      );
     } finally {
       setLoading(false);
     }
   };
+
   const googleProvider = new GoogleAuthProvider();
+
+  const handleNameSubmit = async (name: any) => {
+    if (newUser) {
+      try {
+        const userDocRef = doc(db, "users", newUser.uid);
+        await setDoc(userDocRef, {
+          email: newUser.email,
+          name: name,
+          uid: newUser.uid,
+        });
+
+        toast.success("Google sign-in successful!", { position: "top-right" });
+        localStorage.setItem("min_tomt_login", "true");
+        localStorage.setItem("Iplot_email", newUser.email);
+        router.push("/");
+      } catch (error) {
+        console.error("Error saving new user", error);
+        toast.error("Registration failed. Please try again.", {
+          position: "top-right",
+        });
+      } finally {
+        setIsModalOpen(false); // Close the modal
+      }
+    }
+  };
 
   const signInWithGoogle = async () => {
     setLoading(true);
     try {
       const result = await signInWithPopup(auth, googleProvider);
       const user: any = result.user;
-      toast.success("Google sign-in successful!", { position: "top-right" });
-      localStorage.setItem("min_tomt_login", "true");
-      localStorage.setItem("Iplot_email", user?.email);
+
+      const userDocRef = doc(db, "users", user.uid);
+      const userDoc = await getDoc(userDocRef);
+
+      if (userDoc.exists()) {
+        toast.success("Google sign-in successfull!", { position: "top-right" });
+        localStorage.setItem("min_tomt_login", "true");
+        localStorage.setItem("Iplot_email", user?.email);
+        router.push("/");
+      } else {
+        setNewUser(user);
+        setIsModalOpen(true);
+      }
     } catch (error) {
       console.error("Error during Google Sign-In", error);
       toast.error("Google sign-in failed. Please try again.", {
@@ -159,6 +202,11 @@ const Login = () => {
           <Loader />
         </div>
       )}
+      <NameModal
+        isOpen={isModalOpen}
+        onClose={() => setIsModalOpen(false)}
+        onSubmit={handleNameSubmit}
+      />
     </div>
   );
 };
